move the plugin models
