+++ src/packages/core/components/input-tiny-mce/index.ts
@@ -3,3 +3,4 @@ export * from './input-tiny-mce.element.js';
 export * from './input-tiny-mce.languages.js';
 export * from './input-tiny-mce.sanitizer.js';
 export * from './input-tiny-mce.handlers.js';
+export * from './tiny-mce-plugin.js';
