+++ src/packages/core/components/input-tiny-mce/tiny-mce-plugin.ts
@@ -1,0 +1,21 @@
+import { UmbLitElement } from '@umbraco-cms/internal/lit-element';
+import type { UmbDataTypePropertyCollection } from '@umbraco-cms/backoffice/components';
+import type { tinymce } from '@umbraco-cms/backoffice/external/tinymce';
+
+export class UmbTinyMcePluginBase {
+	host: UmbLitElement;
+	editor: tinymce.Editor;
+	configuration?: UmbDataTypePropertyCollection;
+
+	constructor(arg: TinyMcePluginArguments) {
+		this.host = arg.host;
+		this.editor = arg.editor;
+		this.configuration = arg.configuration;
+	}
+}
+
+export interface TinyMcePluginArguments {
+	host: UmbLitElement;
+	editor: tinymce.Editor;
+	configuration?: UmbDataTypePropertyCollection;
+}
