+++ src/packages/core/extension-registry/index.ts
@@ -2,4 +2,3 @@ export * from './interfaces/index.js';
 export * from './models/index.js';
 export * from './registry.js';
 export * from './extension-class-initializer.js';
-export * from './tiny-mce-plugin.js';
